+++ builders/laravel.js
@@ -300,6 +300,6 @@ module.exports = {
 
       // Send downstream
       super(id, _.merge({}, config, options));
-    };
+    }
   },
 };